requirements.txt
# Core Dependencies
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
trl>=0.7.4

# Dataset & Data Processing
datasets>=2.15.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Triton for Custom Kernels
triton>=2.1.0

# Inference Optimization
vllm>=0.2.6
xformers>=0.0.23
flash-attn>=2.3.0

# RAG Components
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.18

# API & Interface
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.8.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Monitoring & Visualization
wandb>=0.16.0
tensorboard>=2.15.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
huggingface-hub>=0.19.0
python-dotenv>=1.0.0
tokenizers>=0.15.0

# Evaluation
rouge-score>=0.1.2
sacrebleu>=2.3.1
bert-score>=0.3.13

# Docker/Deployment
gunicorn>=21.2.0
